refactor(core): hoist toMessagePartStatus to the normalizePartStatus leaf util
Keeps the store entry's runtime graph free of the runtime/api classes.
