+++ packages/core/src/runtime/api/message-runtime.ts
@@ -1,17 +1,9 @@
 import type { SpeechState } from "../interfaces/thread-runtime-core";
 import { symbolInnerMessage } from "../utils/external-store-message";
-import type {
-  ToolCallMessagePartStatus,
-  ThreadMessage,
-  ThreadAssistantMessagePart,
-  ThreadUserMessagePart,
-} from "../../types/message";
+import type { ThreadMessage } from "../../types/message";
 import type { Unsubscribe } from "../../types/unsubscribe";
 import type { MessagePartStatus, RunConfig } from "../../types/message";
-import {
-  COMPLETE_STATUS,
-  normalizePartStatus,
-} from "../../utils/normalizePartStatus";
+import { toMessagePartStatus } from "../../utils/normalizePartStatus";
 import { getThreadMessageText } from "../../utils/text";
 import { NestedSubscriptionSubject } from "../../subscribable/subscribable";
 import {
@@ -35,31 +27,6 @@ import {
 import type { MessageRuntimePath } from "./paths";
 import type { ThreadRuntimeCoreBinding } from "./thread-runtime";
 import type { MessageStateBinding } from "./bindings";
-
-export const toMessagePartStatus = (
-  message: ThreadMessage,
-  partIndex: number,
-  part: ThreadUserMessagePart | ThreadAssistantMessagePart,
-): ToolCallMessagePartStatus => {
-  if (message.role !== "assistant") return COMPLETE_STATUS;
-
-  if (part.type === "tool-call") {
-    if (part.result === undefined) {
-      return message.status as ToolCallMessagePartStatus;
-    } else {
-      return COMPLETE_STATUS;
-    }
-  }
-
-  if (message.status.type === "running") {
-    const status = normalizePartStatus(part);
-    if (status) return status;
-  }
-
-  const isLastPart = partIndex === Math.max(0, message.content.length - 1);
-  if (message.status.type === "requires-action") return COMPLETE_STATUS;
-  return isLastPart ? (message.status as MessagePartStatus) : COMPLETE_STATUS;
-};
 
 const getMessagePartState = (
   message: MessageState,
